+++ packages/vendure-plugin-shipping-extensions/src/strategies/order-address-to-geolocation-strategy.ts
@@ -4,7 +4,7 @@ import { OrderAddress } from '@vendure/common/lib/generated-shop-types';
 export interface OrderAddressToGeolocationConversionStrategy {
   getGeoLocationForAddress: (
     orderAddress: OrderAddress
-  ) => Promise<GeoLocation>;
+  ) => Promise<GeoLocation | undefined>;
 }
 
 export interface GeoLocation {