+++ packages/vendure-plugin-shipping-extensions/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 - Made `ShippingExtensionsOptions.weightCalculationFunction` async and accept `Injector` as an argument
 
+# 2.4.1 (2024-02-22)
+
+- Don't log errors when invalid postalcode for UK strategy
+- Distance based strategies are allowed to return an undefined geolocation
+
 # 2.4.0 (2024-02-14)
 
 - Added `orderInCountryPromotionCondition`, which checks if the order's shipping country matches one of the configured countries